:recycle: standardize the form's names and move form in src/Form/Type
