+++ src/Form/Type/FigureType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Type;
 
 use App\Entity\Figure;
 use App\Entity\Group;
@@ -12,9 +12,8 @@ use Symfony\Component\Form\Extension\Core\Type\TextareaType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
-use Symfony\Component\Validator\Constraints\Image;
 
-class FigureForm2 extends AbstractType
+class FigureType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
@@ -45,6 +44,37 @@ class FigureForm2 extends AbstractType
                     "class" => "vertical-checkboxes",
                 ],
             ])
+            ->add('images', FileType::class, [
+                'label' => false,
+                'multiple' => true,
+                'mapped' => false,
+                'required' => false,
+                'attr' =>
+                [
+                    "class" => "d-flex",
+                ]
+//                'constraints' => [
+//                    new Image([
+////                        'maxSize' => '5M',
+//                        'mimeTypesMessage' => 'Merci d\'uploader une image valide (jpeg/png/webp)',
+////                        'minWidth' => 300,
+////                        'minHeight' => 300,
+//                        // d’autres options disponibles si besoin
+//                    ])
+//                ],
+            ])
+            ->add('videoFigures', CollectionType::class, [
+                'entry_type' => VideoFigureType::class,
+                'entry_options' => ['label' => 'URL de la vidéo (YouTube, Dailymotion...)'],
+                'allow_add' => true,
+                'allow_delete' => true,
+                'prototype' => true,
+                'mapped' => false, // on les traite manuellement
+                'required' => false,
+                'by_reference' => false,
+                'label' => false,
+            ])
+
         ;
     }
