+++ src/Form/Type/PictureFigureType.php
@@ -1,16 +1,14 @@
 <?php
 
-namespace App\Form;
+namespace App\Form\Type;
 
-use App\Entity\Figure;
 use App\Entity\PictureFigure;
-use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\FileType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-class PictureFigureFormType extends AbstractType
+class PictureFigureType extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
